+++ config.py
@@ -6,14 +6,14 @@ MEAN = (0.5463, 0.5463, 0.5463)
 STD = (0.2366, 0.2366, 0.2366)
 
 # Model and Data Parameters
-IMG_SIZE = 300
+IMG_SIZE = 224
 EPOCHS = 64
-BATCH_SIZE = 24
-LEARNING_RATE = 0.0001
+BATCH_SZ = 24
+LR = 0.0001
 NUM_CLASSES = 15
-NUM_WORKERS = 8
+NUM_WRKRS = 8
 DROPOUT_RATE = 0.3
-DROPPUT_PATIENCE = 10
+PATIENCE = 8
 
 # Dataset distribution
 RAND_STATE = 42
